creating qbcore.functions in a central file and importing it
