+++ bridge/qb/shared.lua
@@ -1,0 +1,34 @@
+QBSharedCompat = QBShared
+
+---@deprecated use CommaValue from imports/utils.lua
+QBSharedCompat.CommaValue = CommaValue
+
+---@deprecated use RandomLetter from imports/utils.lua
+QBSharedCompat.RandomStr = RandomLetter
+
+---@deprecated use RandomNumber from imports/utils.lua
+QBSharedCompat.RandomInt = RandomNumber
+
+---@deprecated use string.split from imports/utils.lua
+QBSharedCompat.SplitStr = string.split
+
+---@deprecated use string.trim from imports/utils.lua
+QBSharedCompat.Trim = string.trim
+
+---@deprecated use string.firstToUpper from imports/utils.lua
+QBSharedCompat.FirstToUpper = string.firstToUpper
+
+---@deprecated use math.round from imports/utils.lua
+QBSharedCompat.Round = math.round
+
+---@deprecated use ChangeVehicleExtra from imports/utils.lua
+QBSharedCompat.ChangeVehicleExtra = ChangeVehicleExtra
+
+---@deprecated use SetVehicleExtras from imports/utils.lua
+QBSharedCompat.SetDefaultVehicleExtras = SetVehicleExtras
+
+---@deprecated use MaleNoGloves from imports/utils.lua
+QBSharedCompat.MaleNoGloves = MaleNoGloves
+
+---@deprecated use FemaleNoGloves from imports/utils.lua
+QBSharedCompat.FemaleNoGloves = FemaleNoGloves
