+++ server/player.lua
@@ -494,66 +494,6 @@ function QBCore.Player.CreatePlayer(playerData, Offline)
     end
 end
 
----Add a new function to the Functions table of the player class
----Use-case:
--- [[
---     AddEventHandler('QBCore:Server:PlayerLoaded', function(Player)
---         QBCore.Functions.AddPlayerMethod(Player.PlayerData.source, "functionName", function(oneArg, orMore)
---             -- do something here
---         end)
---     end)
--- ]]
----@param ids number|number[] which players to add the method to. -1 for all players
----@param methodName string
----@param handler function
-function QBCore.Functions.AddPlayerMethod(ids, methodName, handler)
-    local idType = type(ids)
-    if idType == "number" then
-        if ids == -1 then
-            for _, v in pairs(QBCore.Players) do
-                v.Functions.AddMethod(methodName, handler)
-            end
-        else
-            if not QBCore.Players[ids] then return end
-
-            QBCore.Players[ids].Functions.AddMethod(methodName, handler)
-        end
-    elseif idType == "table" and table.type(ids) == "array" then
-        for i = 1, #ids do
-            QBCore.Functions.AddPlayerMethod(ids[i], methodName, handler)
-        end
-    end
-end
-
----Add a new field table of the player class
----Use-case:
---[[
-    AddEventHandler('QBCore:Server:PlayerLoaded', function(Player)
-        QBCore.Functions.AddPlayerField(Player.PlayerData.source, "fieldName", "fieldData")
-    end)
-]]
----@param ids number|number[] which players to add a new field to. -1 for all players
----@param fieldName string
----@param data any
-function QBCore.Functions.AddPlayerField(ids, fieldName, data)
-    local idType = type(ids)
-    if idType == "number" then
-        if ids == -1 then
-            for _, v in pairs(QBCore.Players) do
-                v.Functions.AddField(fieldName, data)
-            end
-        else
-            if not QBCore.Players[ids] then return end
-
-            QBCore.Players[ids].Functions.AddField(fieldName, data)
-        end
-    elseif idType == "table" and table.type(ids) == "array" then
-        for i = 1, #ids do
-            QBCore.Functions.AddPlayerField(ids[i], fieldName, data)
-        end
-    end
-end
-
 ---Save player info to database (make sure citizenid is the primary key in your database)
 ---@param source Source
 function QBCore.Player.Save(source)
