+++ shared/main.lua
@@ -1,34 +1,1 @@
 QBShared = QBShared or {}
-
----@deprecated use CommaValue from imports/utils.lua
-QBShared.CommaValue = CommaValue
-
----@deprecated use RandomLetter from imports/utils.lua
-QBShared.RandomStr = RandomLetter
-
----@deprecated use RandomNumber from imports/utils.lua
-QBShared.RandomInt = RandomNumber
-
----@deprecated use string.split from imports/utils.lua
-QBShared.SplitStr = string.split
-
----@deprecated use string.trim from imports/utils.lua
-QBShared.Trim = string.trim
-
----@deprecated use string.firstToUpper from imports/utils.lua
-QBShared.FirstToUpper = string.firstToUpper
-
----@deprecated use math.round from imports/utils.lua
-QBShared.Round = math.round
-
----@deprecated use ChangeVehicleExtra from imports/utils.lua
-QBShared.ChangeVehicleExtra = ChangeVehicleExtra
-
----@deprecated use SetVehicleExtras from imports/utils.lua
-QBShared.SetDefaultVehicleExtras = SetVehicleExtras
-
----@deprecated use MaleNoGloves from imports/utils.lua
-QBShared.MaleNoGloves = MaleNoGloves
-
----@deprecated use FemaleNoGloves from imports/utils.lua
-QBShared.FemaleNoGloves = FemaleNoGloves
